+++ src/main.cpp
@@ -60,7 +60,7 @@ int main(int argc, char *argv[]) {
         Uint32 current = SDL_GetTicks();
         Uint32 time = SDL_GetTicks();
 
-        float dT = (current - lastUpdate) / 1000.0f;
+        double dT = (current - lastUpdate) / 1000.0f;
         glm::vec3 point_to_seek = glm::vec3(x, y, 0);
 
         flock.update(dT, point_to_seek);
